separate command line functionality from library functionality so it can work as both
