+++ package.json
@@ -4,7 +4,7 @@
   "description": "CLI for Status",
   "main": "index.js",
   "bin": {
-    "status-dev-cli": "./index.js"
+    "status-dev-cli": "./cli.js"
   },
   "author": "Status.im",
   "license": "MPL-2.0",
